API Reference

On this page the docs of the provided functions are listed

FerriteViz.solutionplot — Function

solutionplot(plotter::MakiePlotter; kwargs...)
solutionplot(dh::AbstractDofHandler, u::Vector; kwargs...)
solutionplot!(plotter::MakiePlotter; kwargs...)
solutionplot!(dh::AbstractDofHandler, u::Vector; kwargs...)

Solutionplot produces the classical contour plot onto the finite element mesh. Most important keyword arguments are:

field::Symbol=:default representing the field which gets plotted, defaults to the first field in the dh.

deformation_field::Symbol=:default field that transforms the mesh by the given deformation, defaults to no deformation

process::Function=postprocess function to construct nodal scalar values from a vector valued problem

colormap::Symbol=:cividis

deformation_scale=1.0

shading=false

scale_plot=false

transparent=false

FerriteViz.cellplot — Function

cellplot(plotter::MakiePlotter,σ::Vector{T}; kwargs...) where T
cellplot!(plotter::MakiePlotter,σ::Vector{T}; kwargs...) where T

cellplot plots constant scalar data on the cells of the finite element mesh. If T is not a number, the keyword argument process can be passed in order to reduce the elements of σ to a scalar.

keyword arguments are:

deformation_field::Symbol=:default field that transforms the mesh by the given deformation, defaults to no deformation

process::Function=identity function to construct cell scalar values. Defaults to identity, i.e. scalar values.

colormap::Symbol=:cividis

deformation_scale=1.0

shading=false

scale_plot=false

transparent=false

FerriteViz.wireframe — Function

wireframe(plotter::MakiePlotter; kwargs...)
wireframe(dh::AbstractDofHandler, u::Vector; kwargs...)
wireframe(grid::AbstractGrid; kwargs...)
wireframe!(plotter::MakiePlotter; kwargs...)
wireframe!(dh::AbstractDofHandler, u::Vector; kwargs...)
wireframe!(grid::AbstractGrid; kwargs...)

Plots the finite element mesh, optionally labels it and transforms it if a suitable deformation_field is given.

plotnodes::Bool=true plots the nodes as circles/spheres

strokewidth::Int=2 how thick faces/edges are drawn

color::Symbol=theme(scene,:linecolor) color of the faces/edges and nodes

markersize::Int=30 size of the nodes

deformation_field::Symbol=:default field that transforms the mesh by the given deformation, defaults to no deformation

deformation_scale::Number=1.0 scaling of the deformation

cellsets=false Color cells based on their cellset association. If no cellset is found for a cell, the cell is marked blue.

nodelables=false global node id labels

nodelabelcolor=:darkblue

celllabels=false global cell id labels

celllabelcolor=:darkred

textsize::Int=15 size of the label's text

visible=true

FerriteViz.arrows — Function

arrows(plotter::MakiePlotter; kwargs...)
arrows(dh::AbstractDofHandler, u::Vector; kwargs...)
arrows!(plotter::MakiePlotter; kwargs...)
arrows!(dh::AbstractDofHandler, u::Vector; kwargs...)

At every node position a arrows is drawn, where the arrow tip ends at the node. Only works in dim >=2. If a color is specified the arrows are unicolored. Otherwise the color corresponds to the magnitude, or any other scalar value based on the process function.

arrowsize = 0.08

normalize = true

field = :default

color = :default

colormap = :cividis

process=postprocess

lengthscale = 1f0

FerriteViz.surface — Function

surface(plotter::MakiePlotter; kwargs...)
surface(dh::AbstractDofHandler, u::Vector; kwargs...)
surface!(plotter::MakiePlotter; kwargs...)
surface!(dh::AbstractDofHandler, u::Vector; kwargs...)

Uses the given field and plots the scalar values as a surface. If it's a vector valued problem, the nodal vector values are transformed to a scalar based on process which defaults to the magnitude. Only availble in dim=2.

field = :default

process = postprocess

scale_plot = false

shading = false

colormap = :cividis

FerriteViz.ferriteviewer — Function

ferriteviewer(plotter::MakiePlotter)
ferriteviewer(plotter::MakiePlotter, u_history::Vector{Vector{T}}})

Constructs a viewer with a solutionplot, Colorbar as well as sliders,toggles and menus to change the current view. If the second dispatch is called a timeslider is added, in order to step through a set of solutions obtained from a simulation.

FerriteViz.update! — Function

FerriteViz.update!(plotter::MakiePlotter, u::Vector)

Updates the Observable plotter.u and thereby, triggers the plot to update.